Homepage building
I have been learning about Django shells, URL mapping to views, what
views are, templates and so forth. This has been an interesting few
hours to say the least. I have been interested in the structure of these
applications and how we have link so many moving parts just so that the
we can do something so easy as connecting an .html file to a url instead
of the default 1.800.0.0.1 address given.

09/08/2020